# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/courts/{id}": {
      "get": {
        "operationId": "CourtsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuadraResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "courts"
        ]
      },
      "patch": {
        "operationId": "CourtsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourtDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuadraResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "courts"
        ]
      }
    },
...
    "/api/v1/courts/{id}/imagem": {
      "put": {
        "operationId": "CourtImageController_substituir",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "arquivo",
                  "semPessoasIdentificaveis"
                ],
                "properties": {
                  "arquivo": {
                    "type": "string",
                    "format": "binary"
                  },
                  "semPessoasIdentificaveis": {
                    "type": "string",
                    "enum": [
                      "true"
                    ],
                    "description": "AC-007 — afirmação de que a imagem não mostra pessoas identificáveis. Sem ela, 422 CONFIRMACAO_OBRIGATORIA e nada é gravado."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImagemDaQuadraResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "courts"
        ]
      },
      "delete": {
        "operationId": "CourtImageController_remover",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImagemDaQuadraResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "courts"
        ]
      }
    },
    "/api/v1/court-sports": {
      "get": {
        "operationId": "CourtSportsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CatalogoDeQuadraResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "court-sports"
        ]
      },
      "post": {
        "operationId": "CourtSportsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogoDeQuadraDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogoDeQuadraResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "court-sports"
        ]
      }
    },
    "/api/v1/court-sports/{id}": {
      "patch": {
        "operationId": "CourtSportsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogoDeQuadraDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogoDeQuadraResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "court-sports"
        ]
      },
      "delete": {
        "operationId": "CourtSportsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "court-sports"
        ]
      }
    },
    "/api/v1/court-categories": {
      "get": {
        "operationId": "CourtCategoriesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CatalogoDeQuadraResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "court-categories"
        ]
      },
      "post": {
        "operationId": "CourtCategoriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogoDeQuadraDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogoDeQuadraResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "court-categories"
        ]
      }
    },
    "/api/v1/court-categories/{id}": {
      "patch": {
        "operationId": "CourtCategoriesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogoDeQuadraDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogoDeQuadraResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "court-categories"
        ]
      },
      "delete": {
        "operationId": "CourtCategoriesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "court-categories"
        ]
      }
    },
...
  },
  "info": {
    "title": "PlayCK API",
    "description": "Contrato da API do back — fonte para geração de tipos nos 3 frontends (ADR-001)",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "OpcaoDeCatalogoResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nome": {
            "type": "string",
            "example": "Saibro"
          }
        },
        "required": [
          "id",
          "nome"
        ]
      },
      "QuadraResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "companyId": {
            "type": "string",
            "format": "uuid"
          },
          "nome": {
            "type": "string",
            "example": "Quadra 1"
          },
          "esporte": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/OpcaoDeCatalogoResponseDto"
              }
            ]
          },
          "categoria": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/OpcaoDeCatalogoResponseDto"
              }
            ]
          },
          "precoHora": {
            "type": "number",
            "example": 120
          },
          "status": {
            "type": "string",
            "enum": [
              "ativa",
              "inativa"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "imagemUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "companyId",
          "nome",
          "esporte",
          "categoria",
          "precoHora",
          "status",
          "createdAt",
          "imagemUrl"
        ]
      },
...
      "UpdateCourtDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string"
          },
          "esporteId": {
            "type": "string",
            "format": "uuid"
          },
          "categoriaId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "precoHora": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "ativa",
              "inativa"
            ]
          }
        }
      },
...
      "ImagemDaQuadraResponseDto": {
        "type": "object",
        "properties": {
          "imagemUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "imagemUrl"
        ]
      },
      "CatalogoDeQuadraResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "companyId": {
            "type": "string",
            "format": "uuid"
          },
          "nome": {
            "type": "string",
            "example": "Saibro"
          },
          "ordem": {
            "type": "number",
            "example": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "companyId",
          "nome",
          "ordem",
          "createdAt"
        ]
      },
      "CatalogoDeQuadraDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "example": "Saibro"
          },
          "ordem": {
            "type": "number",
            "example": 0,
            "description": "Ordena na tela. Default 0."
          }
        }
      },
...
    }
  }
}